12_Agent_Containerisation/01-single-agent/requirements.txt
langchain==0.3.7
langchain-openai==0.2.8
python-dotenv==1.0.1
